backend/pyproject.toml
[project]
name = "facesnap"
version = "2.0.0"
description = "统一的人脸检测、识别和人员信息查询API服务"
readme = "../README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "facenet-pytorch>=2.5.3",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
]

# PyTorch CUDA 12.6 索引配置
[tool.uv.sources]
torch = [
    { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
    { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

